Player se movimentando de forma correta!!!! Depois de muitas horas!!!

+++ include/objects.h
@@ -13,7 +13,7 @@ typedef struct object {
 	Object_type  obj_type;
 	SDL_Texture* sprite;
 	int          state;
-	int		     pos_x, pos_y;
+	float		 pos_x, pos_y;
 	SDL_Rect	 collision;	//dst_rect:  pos_x pos_y, width_height
 	
 	struct object* prox;
